server/requirements.txt
flask
flask-mysqldb
flask-cors
werkzeug
sqlalchemy
mysqlclient
joblib
scikit-learn
